# students/k3339/Gneushev_Vladislav/lab3/docs/swagger.json
{
   "openapi":"3.1.0",
   "info":{
      "title":"FastAPI",
      "version":"0.1.0"
   },
   "paths":{
...
      "/api/buses":{
         "get":{
            "tags":[
               "Автобусы"
            ],
            "summary":"Получить список автобусов",
            "operationId":"get_buses_api_buses_get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/GetBusSchema"
                           },
                           "title":"Response Get Buses Api Buses Get"
                        }
                     }
                  }
               }
            }
         },
         "post":{
            "tags":[
               "Автобусы"
            ],
            "summary":"Добавить автобус",
            "operationId":"add_bus_api_buses_post",
            "parameters":[
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AddBusSchema"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetBusSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/buses/{bus_id}":{
         "delete":{
            "tags":[
               "Автобусы"
            ],
            "summary":"Удалить автобус",
            "operationId":"delete_bus_api_buses__bus_id__delete",
            "parameters":[
               {
                  "name":"bus_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Bus Id"
                  }
               },
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/buses/types":{
         "post":{
            "tags":[
               "Автобусы"
            ],
            "summary":"Добавить тип автобуса",
            "operationId":"add_bus_type_api_buses_types_post",
            "parameters":[
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AddBusTypeSchema"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetBusTypeSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get":{
            "tags":[
               "Автобусы"
            ],
            "summary":"Получить список типов автобусов",
            "operationId":"get_bus_types_api_buses_types_get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/GetBusTypeSchema"
                           },
                           "title":"Response Get Bus Types Api Buses Types Get"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/buses/types/{bus_type_id}":{
         "delete":{
            "tags":[
               "Автобусы"
            ],
            "summary":"Удалить тип автобуса",
            "operationId":"delete_bus_type_api_buses_types__bus_type_id__delete",
            "parameters":[
               {
                  "name":"bus_type_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Bus Type Id"
                  }
               },
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components":{
      "schemas":{
...
         "AddBusSchema":{
            "properties":{
               "state_number":{
                  "type":"string",
                  "title":"State Number"
               },
               "bus_type_id":{
                  "type":"integer",
                  "title":"Bus Type Id"
               }
            },
            "type":"object",
            "required":[
               "state_number",
               "bus_type_id"
            ],
            "title":"AddBusSchema"
         },
         "AddBusTypeSchema":{
            "properties":{
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "people_capacity":{
                  "type":"integer",
                  "title":"People Capacity"
               }
            },
            "type":"object",
            "required":[
               "name",
               "people_capacity"
            ],
            "title":"AddBusTypeSchema"
         },
...
         "GetBusSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "bus_type":{
                  "$ref":"#/components/schemas/GetBusTypeSchema"
               },
               "state_number":{
                  "type":"string",
                  "title":"State Number"
               }
            },
            "type":"object",
            "required":[
               "id",
               "bus_type",
               "state_number"
            ],
            "title":"GetBusSchema"
         },
         "GetBusTypeSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "people_capacity":{
                  "type":"integer",
                  "title":"People Capacity"
               }
            },
            "type":"object",
            "required":[
               "id",
               "name",
               "people_capacity"
            ],
            "title":"GetBusTypeSchema"
         },
...
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         },
...
         "ValidationError":{
            "properties":{
               "loc":{
                  "items":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"integer"
                        }
                     ]
                  },
                  "type":"array",
                  "title":"Location"
               },
               "msg":{
                  "type":"string",
                  "title":"Message"
               },
               "type":{
                  "type":"string",
                  "title":"Error Type"
               }
            },
            "type":"object",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "title":"ValidationError"
         }
      }
   }
}